types and interfaces moved to /types.ts
