+++ web-interface/components/CommentContainer.tsx
@@ -1,19 +1,13 @@
-import { comment } from '../components/OpenedActivityCard';
 import Comment from './Comment';
 import openedCard from '../styles/OpenedActivityCard.module.css';
 import {useRef, useEffect} from 'react';
+import { CommentContainerProps } from '../types/types';
 
-export type CommentContainerProps = {
-    "commentsArray" : Array<comment | null>,
-    "justSent" : string,
-    "arrowDown" : any,
-    "color" : string
-}
+
 
 const CommentContainer = ({commentsArray, justSent, arrowDown, color} : CommentContainerProps) =>{
     const bottomRef = useRef<null | HTMLDivElement>(null);
 
-    // scrolls to the bottom of the component
     useEffect(()=>{
         bottomRef.current?.scrollIntoView({behavior: 'smooth'});
     }, [justSent, arrowDown])
